requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
tensorflow>=2.13.0
numpy>=1.26.0
Pillow>=10.3.0
opencv-python-headless>=4.9.0
